requirements.txt
alembic==1.3.2
asgiref==3.2.3
Babel==2.7.0
blinker==1.4
cffi==1.13.2
cryptography==2.8
Django==3.0.1
django-sorcery==0.10.2
ecdsa==0.14.1
importlib-metadata==1.3.0
inflect==4.0.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.0.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.19
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.0
sqlparse==0.3.0
zipp==0.6.0
